cast

+++ src/Resources.cpp
@@ -731,7 +731,9 @@ void Resources::recordGraphicsCommandBuffer(VkCommandBuffer commandBuffer,
   vkCmdBindVertexBuffers(commandBuffer, 0, 1, vertexBuffers1, offsets);
   vkCmdBindIndexBuffer(commandBuffer, indexBufferLandscape, 0,
                        VK_INDEX_TYPE_UINT32);
-  vkCmdDrawIndexed(commandBuffer, world.landscapeIndices.size(), 1, 0, 0, 0);
+  vkCmdDrawIndexed(commandBuffer,
+                   static_cast<uint32_t>(world.landscapeIndices.size()), 1, 0,
+                   0, 0);
 
   // Pipeline 3
   vkCmdBindPipeline(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS,
